% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cluster_sankey.R
\name{cluster_sankey}
\alias{cluster_sankey}
\title{Generate Sankey cluster-resolution plot.}
\usage{
cluster_sankey(seurat, cluster_prefix, min_transition = 500)
}
\arguments{
\item{seurat}{(Seurat): Seurat object.}

\item{cluster_prefix}{(str): Prefix of the cluster column in the Seurat object.}

\item{min_transition}{(numeric): Min. number of cells in cluster -> cluster transition to include in the plot.}
}
\value{
(plotly): Plotly Sankey plot.
}
\description{
Generates a Sankey plot to visualize the size and identity of clusters across different resolutions.
}
\examples{
\dontrun{
cluster_sankey(seurat, cluster_prefix = "SCT_snn_res.", min_transition = 1000)
}

}
