% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_featureplot.R
\name{plot_features}
\alias{plot_features}
\title{Generates an overview of common scRNA QC-features.}
\usage{
plot_features(seurat)
}
\arguments{
\item{seurat}{(Seurat): Seurat object.}
}
\value{
(ggplot2): ggplot2 object.
}
\description{
This generates figures for the following QC-features:
- Relative mitochondrial expression.
- Number of expressed genes.
- Number of UMI.

Within the expression plot, the mean and mean + 3 * sd are indicated.
}
\examples{
1 + 1

}
