% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_kneeplot.R
\name{plot_kneeplot}
\alias{plot_kneeplot}
\title{Generates knee-plot for one or multiple samples based on STARSolo.}
\usage{
plot_kneeplot(folder, samples, umi_threshold = NULL)
}
\arguments{
\item{folder}{(character): Folder containing the results of STAR / STARSolo.}

\item{samples}{(character): Samples to import (will search for files).}

\item{umi_threshold}{(numeric): Custom threshold to visualize next to STARSolo threshold.}
}
\value{
(list): List of ggplot2 objects.
}
\description{
Imports the No. of UMI per barcode and generates a knee-plot.
The selected threshol of STARSolo will be visualized, an alternative threshold
can be specified to visualize this alternative compared to current threshold.
}
\examples{
1 + 1

}
