% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_screeplot.R
\name{plot_screeplot}
\alias{plot_screeplot}
\title{Generates scree-plot for monocle3 object.}
\usage{
plot_screeplot(cds, pc_total = 75, pc_chosen = NULL, pc_threshold = 0.1)
}
\arguments{
\item{cds}{(cell_data_set): cell_data_set with preprocess_cds performed.}

\item{pc_total}{(integer): Total principal components to plot.}

\item{pc_chosen}{(integer): Principal component to draw a horizontal line.}

\item{pc_threshold}{(numeric): Custom threshold for min. variance explained (vertical line).}
}
\value{
(ggplot2): ggplot2 object.
}
\description{
Generates scree-plot for monocle3 object.
}
\examples{
1 + 1

}
