% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_samples_monocle.R
\name{import_samples_monocle}
\alias{import_samples_monocle}
\title{import_samples_monocle}
\usage{
import_samples_monocle(samples, gtf)
}
\arguments{
\item{samples}{(character): Path(s) to STARSolo output folders (filtered).}

\item{gtf}{(character): Path to GTF file used during alignment and for annotation purposes.}
}
\value{
(cell_data_set): Samples and annotations in monocle3 object.
}
\description{
Import one or multiple sci-RNA-seqv3 samples into a monocle3 object.
}
\examples{
1 + 1

}
