% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_import_samples_monocle.R
\name{import_samples_monocle}
\alias{import_samples_monocle}
\title{Import STARSolo results into monocle3.}
\usage{
import_samples_monocle(folder, samples, gtf, metadata = NULL)
}
\arguments{
\item{folder}{(character): Folder containing the results of STAR / STARSolo (GeneFull/filtered).}

\item{samples}{(character): Samples to import (will search for files).}

\item{gtf}{(character): Path to GTF file used during alignment and for annotation purposes.}

\item{metadata}{(tibble): Additional metadata to add to cell_data_set. Should contain a sample_name column.}
}
\value{
(cell_data_set): Cell_data_set containing all combined samples and annotations without further processing.
}
\description{
Import one or multiple sci-RNA-seqv3 STARSolo (filtered) matrices into
a combined cds object with correct gene-annotations.
}
\examples{
1 + 1

}
