% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_featureplot.R
\name{plot_features}
\alias{plot_features}
\title{Generates an overview of common scRNA QC-features.}
\usage{
plot_features(cds)
}
\arguments{
\item{cds}{(character): Folder containing the results of STAR / STARSolo.}
}
\value{
(ggplot2): ggplot2 object.
}
\description{
Generates an overview of common scRNA QC-features.
}
\examples{
1 + 1

}
